Remove bower configuration
It's just causing confusion by being here. There's no doc telling devs
how to use it, bower is old now and not recommended by the developers
and the version of jquery ui installed in the vendors dir wasn't even
the same one as specified in the bower config, so, there's no point in
keeping it really.