Remove scss, use less instead
One file of scss, didn't even need to change anything in it.

The driver for this was a dependency security issue in node-sass
which has since been deprecated, so rather than switch to the new
sass module (tried it, it broke) might as well just move to something
that definitely works and we use in other modules.