refactor: implement the correct interface methods
BREAKING CHANGE: implement the correct interface methods

None of the interface methods (info, can_view, thumbnail, image) actually seem to be used anywhere, while two methods that are called elsewhere (image_info and get_images) were never actually defined on the interface. This removes the unused ones and adds the two used ones. Theoretically, this shouldn't break anything because none of this was doing anything in the first place (if it were, we'd be seeing errors), but I've made it a breaking change just in case.